# packages/contracts/openapi/vinops.openapi.yaml
openapi: 3.1.0
info:
  title: VinOps API - Core Vertical Slice
  version: 1.1.0
  license:
    name: Proprietary - VinOps internal project artifact
    identifier: LicenseRef-Proprietary
  description: >-
    Executable contract for the VIN-MEGA-001 platform foundation and VIN-MEGA-002 Document Control vertical slice.
    Tenant, project, action and resource scope are derived and verified server-side; client identifiers never grant
    access. Current revision is an explicit context pointer changed only by an atomic authorized publication.
  x-vinops-cr-002:
    effective_date: '2026-07-30'
    access_token:
      default_ttl_minutes: 15
      storage: browser-memory-only
      cache_control: no-store
    refresh_credential:
      storage: hashed-server-representation
      rotation: every-refresh
      reuse_response: revoke-session-family-and-emit-security-audit-signal
      default_idle_ttl_days: 7
      default_absolute_ttl_days: 30
    csrf:
      required_for: cookie-authenticated-mutations
      token_binding: session-and-browser-memory
      origin_check: exact-allowlist-with-referer-fallback
    cors:
      origins: exact-environment-allowlist
      credentials_with_wildcard: forbidden
      production_localhost_or_null_default: forbidden
  x-vinops-response-invariants:
    correlation_id: Every response, including errors and 204 responses, returns X-Correlation-ID and any problem body repeats it as correlation_id.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}:
    get:
      operationId: getProject
      tags:
        - Projects
      summary: Return one project only when it is visible through effective tenant and project membership
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project detail visible through the caller's effective tenant and project membership
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotVisible'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
  /projects/{projectId}/rfis:
    get:
      operationId: listRfiRequests
      tags:
        - RFI
      summary: List RFIs for a project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: status
          in: query
          schema:
            type: string
        - name: ball_in_court_organization_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfiRequestList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createRfiRequest
      tags:
        - RFI
      summary: Create an RFI
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RfiRequestCreate'
      responses:
        '201':
          description: Created
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfiRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}/rfis/{rfiId}:
    get:
      operationId: getRfiRequest
      tags:
        - RFI
      summary: Get RFI detail with answers and documents
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/RfiId'
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfiRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotVisible'
  /projects/{projectId}/rfis/{rfiId}/transitions:
    post:
      operationId: transitionRfiRequest
      tags:
        - RFI
      summary: Transition RFI state
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/RfiId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RfiTransition'
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfiRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectId}/rfis/{rfiId}/responses:
    post:
      operationId: createRfiResponse
      tags:
        - RFI
      summary: Submit clarification or official response with revised document attachment
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/RfiId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RfiResponseCreate'
      responses:
        '201':
          description: Created
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}/submittals:
    get:
      operationId: listSubmittals
      tags:
        - Submittals
      summary: List submittals for a project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: status
          in: query
          schema:
            type: string
        - name: submittal_type
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmittalList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createSubmittal
      tags:
        - Submittals
      summary: Create a submittal
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmittalCreate'
      responses:
        '201':
          description: Created
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submittal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}/submittals/{submittalId}:
    get:
      operationId: getSubmittal
      tags:
        - Submittals
      summary: Get submittal details with items and review timeline
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SubmittalId'
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submittal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotVisible'
  /projects/{projectId}/submittals/{submittalId}/transitions:
    post:
      operationId: transitionSubmittal
      tags:
        - Submittals
      summary: Transition submittal state
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SubmittalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmittalTransition'
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submittal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectId}/submittals/{submittalId}/reviews:
    post:
      operationId: createSubmittalReview
      tags:
        - Submittals
      summary: Submit a review decision in maker-checker workflow
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SubmittalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmittalReviewCreate'
      responses:
        '201':
          description: Created review
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmittalReview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token stored in memory only.
    refreshCookie:
      type: apiKey
      in: cookie
      name: vinops_refresh
      description: Opaque rotated credential. Cookie is Secure, HttpOnly, SameSite=Lax, host-only, Path=/api/v1/auth; idle TTL defaults to 7 days and absolute TTL to 30 days. Never log its value.
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RfiId:
      name: rfiId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SubmittalId:
      name: submittalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
  headers:
    CorrelationId:
      description: End-to-end request correlation identifier returned on every response, including errors and 204 responses.
      schema:
        type: string
        format: uuid
...
    RetryAfter:
      description: Delay in seconds before a safe or idempotency-protected retry.
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Malformed request
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or session revoked
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Visible resource but action, state, scope or separation-of-duty denied
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: State, concurrency or idempotency conflict
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ResourceNotVisible:
      description: Resource does not exist or must not be revealed across tenant, membership or visibility boundaries
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: RESOURCE_NOT_VISIBLE
            message_key: errors.resourceNotVisible
            correlation_id: 123e4567-e89b-42d3-a456-426614174000
            retryable: false
    DependencyUnavailable:
      description: A required dependency is unavailable; clients must not blindly retry non-idempotent mutations
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: DEPENDENCY_UNAVAILABLE
            message_key: errors.dependencyUnavailable
            correlation_id: 123e4567-e89b-42d3-a456-426614174000
            retryable: true
  schemas:
...
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
...
    RfiRequestCreate:
      type: object
      required:
        - code
        - title
        - question
        - requesting_partner_organization_id
      properties:
        code:
          type: string
        title:
          type: string
        question:
          type: string
        suggested_solution:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
          default: normal
        location_node_id:
          type: string
          format: uuid
        work_node_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        requesting_partner_organization_id:
          type: string
          format: uuid
        responding_partner_organization_id:
          type: string
          format: uuid
        source_issue_id:
          type: string
          format: uuid
        sla_business_days:
          type: integer
          default: 5
    RfiTransition:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum:
            [
              submit,
              start_review,
              request_clarification,
              provide_clarification,
              answer_official,
              close,
              reopen,
            ]
        comment:
          type: string
    RfiResponseCreate:
      type: object
      required:
        - response_type
        - content
      properties:
        response_type:
          type: string
          enum: [clarification_request, clarification_answer, official_answer]
        content:
          type: string
        revised_document_id:
          type: string
          format: uuid
        revised_document_revision_id:
          type: string
          format: uuid
        file_id:
          type: string
          format: uuid
    RfiResponse:
      type: object
      required:
        - id
        - rfi_id
        - response_type
        - content
        - author_user_id
        - created_at
      properties:
        id:
          type: string
          format: uuid
        rfi_id:
          type: string
          format: uuid
        response_type:
          type: string
          enum: [clarification_request, clarification_answer, official_answer]
        content:
          type: string
        author_user_id:
          type: string
          format: uuid
        author_partner_organization_id:
          type: string
          format: uuid
        revised_document_id:
          type: string
          format: uuid
        revised_document_revision_id:
          type: string
          format: uuid
        file_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    RfiRequest:
      type: object
      required:
        - id
        - project_id
        - organization_id
        - code
        - title
        - question
        - status
        - requesting_partner_organization_id
        - version
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        code:
          type: string
        title:
          type: string
        question:
          type: string
        suggested_solution:
          type: string
        status:
          type: string
          enum: [Draft, Submitted, Under Review, Clarification Required, Official Answered, Closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        location_node_id:
          type: string
          format: uuid
        work_node_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        requesting_partner_organization_id:
          type: string
          format: uuid
        responding_partner_organization_id:
          type: string
          format: uuid
        ball_in_court_organization_id:
          type: string
          format: uuid
        source_issue_id:
          type: string
          format: uuid
        sla_business_days:
          type: integer
        submitted_at:
          type: string
          format: date-time
        due_at:
          type: string
          format: date-time
        answered_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        sla_status:
          type: string
          enum: [ok, warning_48h, warning_24h, breached]
        sla_hours_remaining:
          type: number
        version:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        responses:
          type: array
          items:
            $ref: '#/components/schemas/RfiResponse'
    RfiRequestList:
      type: object
      required:
        - items
        - page
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RfiRequest'
        page:
          $ref: '#/components/schemas/PageMeta'
...
    SubmittalCreate:
      type: object
      required:
        - code
        - title
        - submittal_type
        - maker_partner_organization_id
      properties:
        code:
          type: string
        title:
          type: string
        submittal_type:
          type: string
          enum: [material_sample, shop_drawing, method_statement, product_data, other]
        maker_partner_organization_id:
          type: string
          format: uuid
        lead_contractor_partner_organization_id:
          type: string
          format: uuid
        consultant_partner_organization_id:
          type: string
          format: uuid
        location_node_id:
          type: string
          format: uuid
        work_node_id:
          type: string
          format: uuid
        specification_document_id:
          type: string
          format: uuid
        drawing_document_id:
          type: string
          format: uuid
        sla_business_days:
          type: integer
          default: 7
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubmittalItemCreate'
...
    Submittal:
      type: object
      required:
        - id
        - project_id
        - organization_id
        - code
        - title
        - submittal_type
        - status
        - maker_partner_organization_id
        - version
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        code:
          type: string
        title:
          type: string
        submittal_type:
          type: string
          enum: [material_sample, shop_drawing, method_statement, product_data, other]
        status:
          type: string
          enum:
            [
              Draft,
              Submitted,
              Under Review,
              Approved,
              Approved with Comments,
              Revise and Resubmit,
              Rejected,
              Closed,
            ]
        maker_partner_organization_id:
          type: string
          format: uuid
        lead_contractor_partner_organization_id:
          type: string
          format: uuid
        consultant_partner_organization_id:
          type: string
          format: uuid
        ball_in_court_organization_id:
          type: string
          format: uuid
        location_node_id:
          type: string
          format: uuid
        work_node_id:
          type: string
          format: uuid
        specification_document_id:
          type: string
          format: uuid
        drawing_document_id:
          type: string
          format: uuid
        submitted_at:
          type: string
          format: date-time
        due_at:
          type: string
          format: date-time
        sla_business_days:
          type: integer
        sla_status:
          type: string
          enum: [ok, warning_48h, warning_24h, breached]
        version:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubmittalItem'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/SubmittalReview'
    SubmittalList:
      type: object
      required:
        - items
        - page
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Submittal'
        page:
          $ref: '#/components/schemas/PageMeta'
...